Fixed warning.
git-svn-id: file:///home/nick/svnrepo/repo/new_cyberspace/trunk@15561 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ server/ParcelAuction.cpp
@@ -48,7 +48,7 @@ double ParcelAuction::computeAuctionPrice(TimeStamp time) const
 
 	const double cur_time = (double)time.time;
 
-	double last_resume_time = auction_start_time.time; // end time of lock[i-1], or auction start time if i = 0.
+	double last_resume_time = (double)auction_start_time.time; // end time of lock[i-1], or auction start time if i = 0.
 	double sum_unlocked_time = 0;
 	for(size_t i=0; i<auction_locks.size(); ++i)
 	{
